+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+## 1.1.0
+
+* Show "no data" if there are only zeros available for a country.
+* Add daily hospital and ICU beds occupancy.
+* Set different colours to cases, deaths, hospital and ICU beds occupancy.
+* Fix image file names for countries with commas and spaces in their names.
+
 ## 1.0.0
 
 Initial release.